+++ public/CHANGELOG.md
@@ -8,7 +8,12 @@ Although this tries to adhere to "Semantic Versioning", this is a website, not a
 - N/A
 
 ## [1.7.10] 2017-07-06
-- Support for S2Scans
+### Added
+- Support for S2Scans.
+  - Thanks to @Vesnyx (#104).
+
+### Fixed
+- MangaCow works should properly grab the chapter URL if it contains dots now.
 
 ## [1.7.9] 2017-07-05
 ### Added